+++ lib/java/ydoc-server/src/main/java/module-info.java
@@ -4,6 +4,7 @@ module org.enso.ydoc.server {
   requires org.graalvm.polyglot;
   requires org.slf4j;
   requires static org.graalvm.nativeimage;
+  requires static org.enso.jvm.interop;
 
   // need for hosted JVM access
   opens org.enso.ydoc.server;